Refactored views into multiple files
