+++ librarycore/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path, include
 from django.contrib.auth import views as auth_views
-import librarycore.views as views
+from librarycore import views
 
 urlpatterns = [
     path('', views.index, name='index'),
