+++ librarycore/views/__init__.py
@@ -1,0 +1,6 @@
+from .index import *
+from .book import *
+from .instance import *
+from .user import *
+from .author import *
+
