+++ librarycore/views/helpers.py
@@ -1,0 +1,3 @@
+def isUserAdmin(user):
+    return user.groups.filter(name='library_admins').exists()
+
